# shared/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FC27 Random Team & League Manager API",
    "description": "Backend for the FC27 random team draw and friend-league tracker",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/leagues/{leagueId}/players": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "List the league's roster — owner only. Only active players unless includeInactive=true",
        "operationId": "listPlayers",
        "parameters": [
          {
            "name": "leagueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Players"
        ],
        "summary": "Add a player to the league's roster — owner only. 400 if the display name is already used in this league",
        "operationId": "addPlayer",
        "parameters": [
          {
            "name": "leagueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/leagues/{leagueId}/matches": {
      "get": {
        "tags": [
          "Matches"
        ],
        "summary": "List the league's recorded matches, most recent first — owner only",
        "operationId": "listMatches",
        "parameters": [
          {
            "name": "leagueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Matches"
        ],
        "summary": "Register a final score between two of the league's players (the wheel's draw) — owner only",
        "operationId": "recordMatch",
        "parameters": [
          {
            "name": "leagueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/leagues/{leagueId}/seasons": {
      "get": {
        "tags": [
          "Seasons"
        ],
        "summary": "List the league's past (closed) seasons, most recently closed first — owner only",
        "operationId": "listPastSeasons",
        "parameters": [
          {
            "name": "leagueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeasonResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/leagues/{leagueId}/seasons/{seasonId}/stats": {
      "get": {
        "tags": [
          "Seasons"
        ],
        "summary": "Golden Boot / Best Defense / head-to-head for one past season — owner only",
        "operationId": "getPastSeasonStats",
        "parameters": [
          {
            "name": "leagueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "seasonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueStatsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/leagues/{leagueId}/seasons/{seasonId}/standings": {
      "get": {
        "tags": [
          "Seasons"
        ],
        "summary": "Standings for one past season, frozen at the point it was closed — owner only",
        "operationId": "getPastSeasonStandings",
        "parameters": [
          {
            "name": "leagueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "seasonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StandingRow"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/leagues/{leagueId}/seasons/{seasonId}/matches": {
      "get": {
        "tags": [
          "Seasons"
        ],
        "summary": "Match history for one past season — owner only",
        "operationId": "getPastSeasonMatches",
        "parameters": [
          {
            "name": "leagueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "seasonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/leagues/{leagueId}/players/{playerId}": {
      "delete": {
        "tags": [
          "Players"
        ],
        "summary": "Remove a player from the roster (soft — past matches keep their record) — owner only",
        "operationId": "removePlayer",
        "parameters": [
          {
            "name": "leagueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PlayerCreateRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PlayerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "leagueId": {
            "type": "integer",
            "format": "int64"
          },
          "displayName": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "MatchCreateRequest": {
        "type": "object",
        "properties": {
          "player1Id": {
            "type": "integer",
            "format": "int64"
          },
          "player2Id": {
            "type": "integer",
            "format": "int64"
          },
          "team1Id": {
            "type": "integer",
            "format": "int64"
          },
          "team2Id": {
            "type": "integer",
            "format": "int64"
          },
          "score1": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "score2": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "required": [
          "player1Id",
          "player2Id",
          "score1",
          "score2",
          "team1Id",
          "team2Id"
        ]
      },
      "MatchResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "leagueId": {
            "type": "integer",
            "format": "int64"
          },
          "player1Id": {
            "type": "integer",
            "format": "int64"
          },
          "player1Name": {
            "type": "string"
          },
          "player2Id": {
            "type": "integer",
            "format": "int64"
          },
          "player2Name": {
            "type": "string"
          },
          "team1Id": {
            "type": "integer",
            "format": "int64"
          },
          "team1Name": {
            "type": "string"
          },
          "team2Id": {
            "type": "integer",
            "format": "int64"
          },
          "team2Name": {
            "type": "string"
          },
          "score1": {
            "type": "integer",
            "format": "int32"
          },
          "score2": {
            "type": "integer",
            "format": "int32"
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "HeadToHeadRow": {
        "type": "object",
        "properties": {
          "player1Id": {
            "type": "integer",
            "format": "int64"
          },
          "player1Name": {
            "type": "string"
          },
          "player2Id": {
            "type": "integer",
            "format": "int64"
          },
          "player2Name": {
            "type": "string"
          },
          "player1Wins": {
            "type": "integer",
            "format": "int32"
          },
          "player2Wins": {
            "type": "integer",
            "format": "int32"
          },
          "draws": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LeagueStatsResponse": {
        "type": "object",
        "properties": {
          "topScorer": {
            "$ref": "#/components/schemas/StatLeader"
          },
          "bestDefense": {
            "$ref": "#/components/schemas/StatLeader"
          },
          "headToHead": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeadToHeadRow"
            }
          }
        }
      },
      "StatLeader": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "integer",
            "format": "int64"
          },
          "displayName": {
            "type": "string"
          },
          "value": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "StandingRow": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "integer",
            "format": "int64"
          },
          "displayName": {
            "type": "string"
          },
          "played": {
            "type": "integer",
            "format": "int32"
          },
          "won": {
            "type": "integer",
            "format": "int32"
          },
          "drawn": {
            "type": "integer",
            "format": "int32"
          },
          "lost": {
            "type": "integer",
            "format": "int32"
          },
          "goalsFor": {
            "type": "integer",
            "format": "int32"
          },
          "goalsAgainst": {
            "type": "integer",
            "format": "int32"
          },
          "goalDifference": {
            "type": "integer",
            "format": "int32"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SeasonResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}